working SPT
Will get a working model online next for simple download.

This is mainly an implementation of TrackPy's Crocker Grier Algorithm implmentation without some of the numba uses, as well as optimization for a potential screening tool for having 0 particles present. This shouldnt reply with an empty dataframe, which was a big problem when using trackpy.